requirements.txt
altgraph==0.17.4
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
et-xmlfile==1.1.0
idna==3.8
jdcal==1.4.1
macholib==1.16.3
openpyxl==2.6.2
packaging==24.1
pillow==10.4.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyPDF2==1.26.0
requests==2.32.3
setuptools==73.0.1
soupsieve==2.6
urllib3==2.2.2
